block: add REQ_URGENT to request flags
This patch adds a new flag to be used in cmd_flags field of struct
request for marking request as urgent. Urgent request is the one that
should be given priority currently handled (regular) request by the
device driver. The decision of a request urgency is taken by the
scheduler.

Change-Id: Ic20470987ef23410f1d0324f96f00578f7df8717
Signed-off-by: Tatyana Brokhman <tlinder@codeaurora.org>

Conflicts:
	include/linux/blk_types.h